+++ CHANGELOG.md
@@ -74,4 +74,6 @@ Changes from v0.4 to current version:
 
 * Added ability to truncate Sersic profiles with optional trunc parameter. (Issue #388)
 
+* Added ability to specify Sersic profiles by the scale radius. (Issue #420)
+
 * Added trefoil to optical aberration. (Issue #390)